refactor(backend): 拆分 ServerFacade 为领域服务 + 轻量协调器
ServerFacade 从 1924 行的上帝对象重构为 324 行的薄委托层。

新增共享内核模块：
- service_result.h: 统一服务层 DTO
- request_parsing.{h,cpp}: 请求解析与校验工具
- json_serialization.{h,cpp}: JSON 序列化辅助函数
- colordb_resolution.{h,cpp}: ColorDB 解析与通道键归一化

新增 6 个领域 Service（依赖注入，无跨服务直接调用）：
- ColorDbService / TaskService / CalibrationService
- ConvertService / MattingVectorizeService / RecipeEditorService

ServerFacade 保留跨域查询并委托所有业务方法到子服务。
API 契约不变，纯内部重构。
